Move moderated session evaluation to its own package
In an attempt to reduce the scope of `lib/auth` session access
evaluation has been moved into the `lib/auth/moderation` package.
