+++ lib/auth/moderation/session_access_test.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package auth
+package moderation
 
 import (
 	"testing"
